requirements.txt
# ML Libraries
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
albumentations>=1.3.1
ultralytics>=8.0.0  # For YOLOv8

# MLOps
mlflow>=2.7.0
kubeflow>=1.8
apache-airflow>=2.7.0

# Cloud SDKs
boto3>=1.28.0  # AWS
azure-ai-ml>=1.10.0  # Azure
google-cloud-storage>=2.10.0  # GCP

# Monitoring and Logging
wandb>=0.15.0
tensorboard>=2.13.0

# API and Utils
fastapi>=0.103.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pytest>=7.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0 